requirements.txt
Django>= 4.0
python-decouple
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
gunicorn
Pillow
flake8
